src/requirements.txt
boto3==1.20.19
botocore==1.23.19
certifi==2019.3.9
chardet==3.0.4
charset-normalizer==2.0.6
click==7.1.2
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
idna==2.8
itsdangerous==0.24
Jinja2==2.10
jmespath==0.10.0
MarkupSafe==1.1.1
Pillow==8.4.0
python-dateutil==2.8.2
requests==2.26.0
s3transfer==0.5.0
six==1.16.0
SQLAlchemy==1.3.1
urllib3==1.26.7
Werkzeug==0.14.1
